Hotfix migraciones

+++ database/migrations/2018_07_30_112411_incluye_campo_beca_en_actividades.php
@@ -27,8 +27,10 @@ class IncluyeCampoBecaEnActividades extends Migration
      */
     public function down()
     {
-        Schema::table('Actividad', function (Blueprint $table) {
-            $table->dropColumn('beca');
-        });
+        if (Schema::hasColumn('Actividad', 'beca')) {
+            Schema::table('Actividad', function (Blueprint $table) {
+                $table->dropColumn('beca');
+            });
+        }
     }
 }
